requirements.txt
PySide6>=6.5.0
Pillow>=10.0.0
reportlab>=4.0.0
google-generativeai>=0.3.0
